requirements.txt
matplotlib==3.8.0
plotly.express==5.24.1
numpy==1.25.2
pandas==2.1.4
seaborn==0.13.0
streamlit==1.30.0